특강 목록 api 서비스 및 테스트 리팩토링
