+++ src/test/kotlin/dev/lecture_clean_tdd/application/service/RegisterLectureServiceTest.kt
@@ -8,7 +8,8 @@ import dev.lecture_clean_tdd.Exception.MaxAttendeesReachedException
 import dev.lecture_clean_tdd.Exception.UserNotFoundException
 import dev.lecture_clean_tdd.domain.entity.Lecture
 import dev.lecture_clean_tdd.application.port.output.LectureRepository
-import dev.lecture_clean_tdd.adapter.web.request.LectureRequestDto
+import dev.lecture_clean_tdd.adapter.web.request.LectureRequest
+import dev.lecture_clean_tdd.adapter.web.request.toDto
 import dev.lecture_clean_tdd.application.listener.event.LectureHistoryEvent
 import dev.lecture_clean_tdd.application.port.output.LectureAttendeeRepository
 import dev.lecture_clean_tdd.application.port.output.UserRepository
@@ -48,7 +49,7 @@ class RegisterLectureServiceTest {
     fun `특강을 신청하면 성공한다`() {
         val userId = 1L
         val lectureId = 1L
-        val request = LectureRequestDto(userId, lectureId)
+        val request = LectureRequest(userId, lectureId)
 
         val lecture = Lecture("테스트", "2024-04-10 13:00", "2024-12-31 13:00", "2025-01-01 13:00")
 
@@ -57,7 +58,7 @@ class RegisterLectureServiceTest {
         `when`(userRepository.findById(request.userId)).thenReturn(user)
         `when`(lectureRepository.findByIdWithLock(request.lectureId)).thenReturn(lecture)
 
-        val result = registerLectureService.registerLecture(request)
+        val result = registerLectureService.registerLecture(request.toDto())
 
         assertThat(result).isTrue()
     }
@@ -66,10 +67,10 @@ class RegisterLectureServiceTest {
     fun `특강 신청시 유효하지 않은 회원이면 UseNotFoundException 을 발생시킨다`() {
         val userId = -1L
         val lectureId = 1L
-        val request = LectureRequestDto(userId, lectureId)
+        val request = LectureRequest(userId, lectureId)
 
         assertThatThrownBy {
-            registerLectureService.registerLecture(request)
+            registerLectureService.registerLecture(request.toDto())
         }.isInstanceOf(UserNotFoundException::class.java)
     }
 
@@ -77,14 +78,14 @@ class RegisterLectureServiceTest {
     fun `특강 신청시 유효하지 않은 강의이면 LectureNotFoundException 을 발생시킨다`() {
         val userId = 1L
         val lectureId = -100L
-        val request = LectureRequestDto(userId, lectureId)
+        val request = LectureRequest(userId, lectureId)
 
         val user = User("변주환")
 
         `when`(userRepository.findById(request.userId)).thenReturn(user)
 
         assertThatThrownBy {
-            registerLectureService.registerLecture(request)
+            registerLectureService.registerLecture(request.toDto())
         }.isInstanceOf(LectureNotFoundException::class.java)
     }
 
@@ -92,7 +93,7 @@ class RegisterLectureServiceTest {
     fun `특강신청이 성공하면 특강의 currentAttendees 가 1 증가한다`() {
         val userId = 1L
         val lectureId = 1L
-        val request = LectureRequestDto(userId, lectureId)
+        val request = LectureRequest(userId, lectureId)
 
         val lecture = Lecture("테스트", "2024-04-10 13:00", "2024-12-31 13:00", "2025-01-01 13:00")
         val user = User("변주환")
@@ -100,7 +101,7 @@ class RegisterLectureServiceTest {
         `when`(userRepository.findById(request.userId)).thenReturn(user)
         `when`(lectureRepository.findByIdWithLock(request.lectureId)).thenReturn(lecture)
 
-        registerLectureService.registerLecture(request)
+        registerLectureService.registerLecture(request.toDto())
 
         assertThat(lecture.currentAttendees).isEqualTo(1)
     }
@@ -109,7 +110,7 @@ class RegisterLectureServiceTest {
     fun `특강인원이 30명이상이면 MaxAttendeesReachedException 이 발생한다`() {
         val userId = 1L
         val lectureId = 1L
-        val request = LectureRequestDto(userId, lectureId)
+        val request = LectureRequest(userId, lectureId)
 
         val lecture = Lecture("테스트", "2024-04-10 13:00", "2024-12-31 13:00", "2025-01-01 13:00")
         val user = User("변주환")
@@ -120,7 +121,7 @@ class RegisterLectureServiceTest {
         `when`(lectureRepository.findByIdWithLock(request.lectureId)).thenReturn(lecture)
 
         assertThatThrownBy {
-            registerLectureService.registerLecture(request)
+            registerLectureService.registerLecture(request.toDto())
         }.isInstanceOf(MaxAttendeesReachedException::class.java)
     }
 
@@ -128,7 +129,7 @@ class RegisterLectureServiceTest {
     fun `동일한 userId 로 신청이력을 검색하여 동일한 특강을 신청하면 DuplicateLectureRegistrationException 이 발생한다`() {
         val userId = 1L
         val lectureId = 1L
-        val request = LectureRequestDto(userId, lectureId)
+        val request = LectureRequest(userId, lectureId)
 
         val lecture = Lecture("테스트", "2024-04-10 13:00", "2024-12-31 13:00", "2025-01-01 13:00")
         val user = User("변주환")
@@ -139,7 +140,7 @@ class RegisterLectureServiceTest {
         `when`(lectureAttendeeRepository.findByUserAndLecture(user,lecture)).thenReturn(lectureAttendee)
 
         assertThatThrownBy {
-            registerLectureService.registerLecture(request)
+            registerLectureService.registerLecture(request.toDto())
         }.isInstanceOf(DuplicateLectureRegistrationException::class.java)
             .hasMessage("이미 이 강의에 신청하셨습니다")
     }
@@ -148,7 +149,7 @@ class RegisterLectureServiceTest {
     fun `특강 신청 시각보다 먼저 신청시 EarlyLectureRegistrationException 을 발생 시킨다`() {
         val userId = 1L
         val lectureId = 1L
-        val request = LectureRequestDto(userId, lectureId)
+        val request = LectureRequest(userId, lectureId)
 
         val lecture = Lecture("테스트", "2025-01-24 21:00", "2026-12-31 13:00", "2027-01-01 13:00")
         val user = User("변주환")
@@ -157,7 +158,7 @@ class RegisterLectureServiceTest {
         `when`(lectureRepository.findByIdWithLock(request.lectureId)).thenReturn(lecture)
 
         assertThatThrownBy {
-            registerLectureService.registerLecture(request)
+            registerLectureService.registerLecture(request.toDto())
         }.isInstanceOf(EarlyLectureRegistrationException::class.java)
     }
 
@@ -165,7 +166,7 @@ class RegisterLectureServiceTest {
     fun `특강 신청 종료일보다 늦게 신청시 LateLectureRegistrationException 을 발생 시킨다`() {
         val userId = 1L
         val lectureId = 1L
-        val request = LectureRequestDto(userId, lectureId)
+        val request = LectureRequest(userId, lectureId)
 
         val lecture = Lecture("테스트", "2024-04-10 13:00", "2024-04-10 15:00", "2025-01-01 13:00")
         val user = User("변주환")
@@ -174,7 +175,7 @@ class RegisterLectureServiceTest {
         `when`(lectureRepository.findByIdWithLock(request.lectureId)).thenReturn(lecture)
 
         assertThatThrownBy {
-            registerLectureService.registerLecture(request)
+            registerLectureService.registerLecture(request.toDto())
         }.isInstanceOf(LateLectureRegistrationException::class.java)
     }
 
@@ -183,7 +184,7 @@ class RegisterLectureServiceTest {
         // given
         val userId = 1L
         val lectureId = 1L
-        val request = LectureRequestDto(userId, lectureId)
+        val request = LectureRequest(userId, lectureId)
 
         val lecture = Lecture("테스트", "2024-04-10 13:00", "2024-12-31 15:00", "2025-01-01 13:00")
         val user = User("변주환")
@@ -192,7 +193,7 @@ class RegisterLectureServiceTest {
         `when`(userRepository.findById(request.userId)).thenReturn(user)
         `when`(lectureRepository.findByIdWithLock(request.lectureId)).thenReturn(lecture)
 
-        registerLectureService.registerLecture(request)
+        registerLectureService.registerLecture(request.toDto())
 
         //then
         verify(eventPublisher, times(1)).publishEvent(any(LectureHistoryEvent::class.java))
